add benchmark script
This commit adds a benchmark script that can be run with `npm run
bench`. The output of the script is in a format that is compatible with
benchstat (https://godoc.org/golang.org/x/perf/cmd/benchstat). This
makes it easy to compare performance between different versions of
Calipers. See `bench/index.js` for more details.

The benchmark script currently only benchmarks measurement of PNG
images. Additional PNG fixtures have been added that the benchmark uses.