implement PNG measurement in typescript
This commit begins the transition to typescript in earnest. Measurement
of PNGs has been implemented.

Note that `npm test` runs both the current test suite and the test suite
for the typescript implementation.

This commit also adds a `npm run bench:ts` command for benchmarking the
typescript implementation. Off that bat there is a performance boost by
not reading from the PNG twice like Calipers v2 does (once to detect the
file type and again to determine the dimensions).

    name                  old time/op  new time/op  delta
    Calipers/png-100x60    108µs ± 4%    79µs ± 2%  -26.91%  (p=0.008 n=5+5)
    Calipers/png-123x456   106µs ± 1%    80µs ± 3%  -25.15%  (p=0.008 n=5+5)
    Calipers/png-128x128   105µs ± 1%    80µs ± 8%  -23.87%  (p=0.008 n=5+5)
    Calipers/png-243x42    105µs ± 2%    78µs ± 1%  -25.03%  (p=0.008 n=5+5)
    Calipers/png-256x171   104µs ± 1%    78µs ± 1%  -25.11%  (p=0.008 n=5+5)
    Calipers/png-304x85    105µs ± 1%    79µs ± 1%  -24.51%  (p=0.008 n=5+5)
    Calipers/png-320x200   106µs ± 4%    79µs ± 1%  -25.48%  (p=0.008 n=5+5)
    Calipers/png-70x46     105µs ± 1%    79µs ± 3%  -24.98%  (p=0.008 n=5+5)
    Calipers/png-70x70     104µs ± 1%    77µs ± 2%  -25.99%  (p=0.008 n=5+5)